+++ PHP-Classes/Factory.class.php
@@ -88,6 +88,7 @@ class Factory {
         $output .= "<section class='S4'></section>";
         $output .= "<section class='S5'></section>";
         $output .= "<p id='Copyright'>Copyright &copy; " . date('Y') . " Darnel-K</p>";
+        $output .= "<p id='Version'>V0.18.02 - 2018-10-12</p>";
         // Not created by, affiliated with, or supported by Slack Technologies, Inc.
         return $output;
     }