+++ Resume/index.php
@@ -12,9 +12,8 @@
         $OpenGraph = array(
             "Title" => "Darnel-K | Resume",
             "Type" => "website",
-            "URL" => ($_SERVER['HTTPS'] ? 'https://' : 'http://') . $_SERVER['HTTP_HOST'] . $_SERVER['REQUEST_URI'],
-            "Image" => ($_SERVER['HTTPS'] ? 'https://' : 'http://') . $_SERVER['HTTP_HOST'] . str_replace(' ', '%20', GlobalVars::$OpenGraphImage),
-            "Image:alt" => "Open Graph Site Resume Page Preview Image",
+            "URL" => (GlobalVars::HTTPS() ? 'https://' : 'http://') . $_SERVER['HTTP_HOST'] . $_SERVER['REQUEST_URI'],
+            "Image" => (GlobalVars::HTTPS() ? 'https://' : 'http://') . $_SERVER['HTTP_HOST'] . str_replace(' ', '%20', GlobalVars::$OpenGraphImage),
             "Description" => "Development area of my site",
             "Locale" => "en_GB"
         );