+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 Versioning numbers defined as MAJOR.MINOR.PATCH
 
+## [0.18.2] - 2018-10-12
+
+### Fixed
+
+-   Undefined index 'HTTPS'
+
+### Added
+
+-   Version number & date in footer
+
 ## [0.18.1] - 2018-09-27
 
 ### Removed